+++ Plugins/Creature/Creature.hpp
@@ -101,6 +101,7 @@ private:
     ArgumentStack SetDisarmable                 (ArgumentStack&& args);
     ArgumentStack SetFaction                    (ArgumentStack&& args);
     ArgumentStack GetFaction                    (ArgumentStack&& args);
+    ArgumentStack GetFlatFooted                 (ArgumentStack&& args);
 
     CNWSCreature *creature(ArgumentStack& args);